requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
h11==0.14.0
idna==3.6
lxml==4.9.3
multidict==6.0.4
nest-asyncio==1.5.8
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stem==1.8.2
torrequest==0.1.0
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4